ui: collapse /team and /team/maintenance into tabbed TeamPage
Move TeamDashboard + TeamMaintenance into a single TeamPage at
/g/<grove>/team with three tabs (Status, Sync, Members) driven by
?tab= query param. Status and Sync bodies move verbatim into
pages/Team/{StatusTab,SyncTab}.tsx; MembersTab is a placeholder
filled in Task 7. NotConnectedView (and its inline ConnectForm)
move to pages/Team/NotConnectedView.tsx.

/g/<grove>/team/maintenance redirects to /g/<grove>/team?tab=sync
for legacy bookmarks.
