+++ packages/myco/ui/src/pages/Team/MembersTab.tsx
@@ -1,0 +1,3 @@
+export function MembersTab() {
+  return <div className="text-sm text-on-surface-variant">Members tab coming up in Task 7.</div>;
+}
